Updated to have a click event on the chart to pull out the specific transactionions. This will need a UI to be able to visibly see what transaction. Pulled out the data from the app view and made it it's own script for reusability.
